added roles/authorization for editing/destroy wikis

+++ app/controllers/wikis_controller.rb
@@ -1,4 +1,6 @@
 class WikisController < ApplicationController
+  before_action :require_sign_in
+
   def index
     @wikis = Wiki.all
   end
@@ -12,7 +14,6 @@ class WikisController < ApplicationController
   end
 
   def create
-
     @wiki = Wiki.new
     @wiki.title = params[:wiki][:title]
     @wiki.body = params[:wiki][:body]
@@ -31,7 +32,9 @@ class WikisController < ApplicationController
   end
 
   def update
+
     @wiki = Wiki.find(params[:id])
+    authorize (@wiki), :update?
     @wiki.title = params[:wiki][:title]
     @wiki.body = params[:wiki][:body]
 
@@ -46,13 +49,22 @@ class WikisController < ApplicationController
 
   def destroy
     @wiki = Wiki.find(params[:id])
-
+    authorize (@wiki)
     if @wiki.destroy
       flash[:notice] = "\"#{@wiki.title}\" was deleted successfully."
       redirect_to wikis_path
     else
       flash.now[:alert] = "There was an error deleting the wiki."
       render :show
+    end
+  end
+
+  private
+
+  def require_sign_in
+    unless current_user
+      flash[:alert] = "You must be logged in to do that"
+      redirect_to new_user_session_path
     end
   end
 end
